+++ src/spreading/gpu.jl
@@ -6,7 +6,8 @@
         @Const(points::NTuple{D}),
         @Const(vp::NTuple{C}),
         @Const(pointperm),
-    ) where {C, D}
+        @Const(transform_fold::F),
+    ) where {F <: Function, C, D}
     i = @index(Global, Linear)
 
     j = if pointperm === nothing
@@ -24,7 +25,7 @@
     @assert eltype(first(us)) <: Real # output is a real array (but may actually describe complex data)
     Ns = spread_actual_dims(Z, Ns_real)  # drop the first dimension (of size 2) if Z <: Complex
 
-    indvals = get_inds_vals_gpu(gs, evalmode, points, Ns, j)
+    indvals = get_inds_vals_gpu(transform_fold, gs, evalmode, points, Ns, j)
 
     v⃗ = map(v -> @inbounds(v[j]), vp)
     spread_onto_arrays_gpu!(us, indvals, v⃗, Ns)
@@ -128,18 +129,17 @@ to_real_array(u::AbstractArray{T}) where {T <: Complex} = reinterpret(reshape, r
 # We assume all arrays are already on the GPU.
 function spread_from_points!(
         backend::GPU,
+        transform_fold::F,
         bd::Union{BlockDataGPU, NullBlockData},
         gs,
         evalmode::EvaluationMode,
         us::NTuple{C, AbstractGPUArray},
-        x⃗s::StructVector,
+        xp::NTuple{D, AbstractGPUVector},
         vp_all::NTuple{C, AbstractGPUVector},
-    ) where {C}
+    ) where {F <: Function, C, D}
     # Note: the dimensions of arrays have already been checked via check_nufft_nonuniform_data.
-    Base.require_one_based_indexing(x⃗s)  # this is to make sure that all indices match
+    foreach(Base.require_one_based_indexing, xp)  # this is to make sure that all indices match
     foreach(Base.require_one_based_indexing, vp_all)
-
-    xs_comp = StructArrays.components(x⃗s)
 
     # Reinterpret `us` as real arrays, in case they are complex.
     # This is to avoid current issues with atomic operations on complex data
@@ -150,12 +150,13 @@ function spread_from_points!(
     sort_points = get_sort_points(bd)::StaticBool  # False in the case of NullBlockData
 
     if pointperm !== nothing
-        @assert eachindex(pointperm) == eachindex(x⃗s)
+        @assert eachindex(pointperm) == eachindex(xp[1])
     end
 
     # We use dynamically sized kernels to avoid recompilation, since number of points may
     # change from one call to another.
-    ndrange_points = size(x⃗s)  # iterate through points
+    ndrange_points = size(xp[1])  # iterate through points
+    @assert all(x -> size(x) == ndrange_points, xp)
     groupsize_points = default_workgroupsize(backend, ndrange_points)
 
     if sort_points === True()
@@ -175,7 +176,7 @@ function spread_from_points!(
     if method === :global_memory
         let ndrange = ndrange_points, groupsize = groupsize_points
             kernel! = spread_from_point_naive_kernel!(backend, groupsize)
-            kernel!(us_real, gs, evalmode, xs_comp, vp_sorted, pointperm_; ndrange)
+            kernel!(us_real, gs, evalmode, xp, vp_sorted, pointperm_, transform_fold; ndrange)
         end
     elseif method === :shared_memory
         @assert bd isa BlockDataGPU
@@ -193,8 +194,8 @@ function spread_from_points!(
             ndrange = gpu_shmem_ndrange_from_groupsize(groupsize, ngroups)
             kernel! = spread_from_points_shmem_kernel!(backend, groupsize, ndrange)
             kernel!(
-                us_real, gs, evalmode, xs_comp, vp_sorted, pointperm_, bd.cumulative_npoints_per_block,
-                HalfSupport(M), block_dims, Val(shmem_size), bd.batch_size,
+                us_real, gs, evalmode, xp, vp_sorted, pointperm_, bd.cumulative_npoints_per_block,
+                HalfSupport(M), block_dims, Val(shmem_size), bd.batch_size, transform_fold,
             )
         end
     end
@@ -239,7 +240,8 @@ end
         ::Val{block_dims},
         ::Val{shmem_size},  # this is a bit redundant, but seems to be required for CPU backends (used in tests)
         ::Val{Np},  # batch_size
-    ) where {C, D, T, Z, M, Np, block_dims, shmem_size}
+        @Const(transform_fold::F),
+    ) where {C, D, T, Z, M, Np, block_dims, shmem_size, F <: Function}
 
     @uniform begin
         groupsize = @groupsize()
@@ -313,7 +315,7 @@ end
                     @inbounds pointperm[i]
                 end
                 g = gs[d]
-                x = points[d][j]
+                x = transform_fold(points[d][j])
                 vp_sm[p] = vp[c][j]
                 gdata = Kernels.evaluate_kernel(evalmode, g, x)
                 ishift = ishifts_sm[d]